% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRegFits.R
\name{logRegModel}
\alias{logRegModel}
\title{Fit logistic regression (binary OR multinomial) using gradient descent}
\usage{
logRegModel(
  x,
  y,
  family = c("binomial", "multinomial"),
  lambda = 0,
  lr = 0.1,
  nIter = 1000
)
}
\arguments{
\item{x}{Numeric matrix or data frame of predictors dim = (n x p)}

\item{y}{Outcome: can be binary numeric 0/1, factor, or character
If there are only two unique values, a binary model is fit
If there are more than two unique values, a one-vs-rest
multinomial model!}

\item{family}{Either "binomial" or "multinomial". If missing,
it is inferred from y!}

\item{lambda}{Greater than 0 L2 penalty on coefficients}

\item{lr}{Learning rate for gradient descent}

\item{nIter}{Number of gradient steps (i.e. iterations)}
}
\value{
An object of class \code{"logRegModel"} with components:
\itemize{
\item beta: coefficient vector (binary) or matrix (multinomial)
\item family: "binomial" or "multinomial"
\item levels: class labels for multinomial
\item lambda, lr, nIter: tuning parameters
}
}
\description{
Fit logistic regression (binary OR multinomial) using gradient descent
}
\examples{
set.seed(1)
x <- matrix(rnorm(200), ncol = 2)
yProb <- 1 / (1 + exp(-(x[,1] - x[,2])))
yBin <- rbinom(nrow(x), size = 1, prob = yProb)
fitBin <- logRegModel(x, yBin, family = "binomial")
pHat <- predict(fitBin, x, type = "response")
mean((pHat > 0.5) == yBin)

# Multinomial example (3 classes)
yMult <- factor(sample(c("A", "B", "C"), nrow(x), replace = TRUE))
fitMulti <- logRegModel(x, yMult, family = "multinomial", nIter = 500)
predMulti <- predict(fitMulti, x, type = "class")
}
