% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRegPreds.R
\name{predict.logRegModel}
\alias{predict.logRegModel}
\title{Predict from a logRegModel object}
\usage{
\method{predict}{logRegModel}(object, newdata, type = c("response", "class"), threshold = 0.5)
}
\arguments{
\item{object}{A \code{"logRegModel"} from \code{logRegModel()}}

\item{newdata}{New predictor data (matrix or data frame)}

\item{type}{"response" for probabilities, "class" for predicted labels}

\item{threshold}{Threshold for binary classification when \code{type = "class"}
@param ... not used, only included for compatibility with \code{predict()}}
}
\value{
A numeric vector, factor, or matrix of probabilities depending on
\code{family} and \code{type}
}
\description{
Predict from a logRegModel object
}
